Add general-purpose compute pool and harden training pipeline
- Add macfleet/compute/ package: TaskSpec, TaskResult, TaskFuture,
  TaskDispatcher (coordinator), TaskWorker (process pool executor)
- Add Pool.map(), Pool.submit(), Pool.run() for arbitrary workloads
- Add `macfleet run` CLI command for script execution
- Add TASK/RESULT wire protocol message types (0x08, 0x09)
- Add cloudpickle dependency for lambda/closure serialization
- Fix 6 engineering issues: move torch out of comm/ (architecture),
  update Engine Protocol to flat numpy, remove DRY violation in
  scheduler, fix torch.load security (weights_only=True), fix
  registry election race condition, remove wasteful float64 promotion
- Add edge case guards: model architecture mismatch detection,
  NaN gradient zeroing, empty gradient skip, node dropout fallback,
  gradient shape validation after allreduce
- Make security default: auto-generate fleet tokens, add --open flag
- 426 tests passing (43 new compute + 383 existing)

+++ macfleet/comm/collectives.py
@@ -174,7 +174,7 @@ class CollectiveGroup:
         # Flatten for chunking
         original_shape = array.shape
         original_dtype = array.dtype
-        flat = array.astype(np.float64).flatten()  # promote for accumulation
+        flat = array.flatten()
         numel = len(flat)
 
         # Pad to be evenly divisible
